creating uiux for login and sginup + changing entity to equal the db

+++ server/src/entities/userRole.entity.ts
@@ -3,7 +3,7 @@ import { User } from './user.entity';
 
 @Entity({ schema: 'skinlab', name: 'user_roles' })
 export class UserRole {
-  @PrimaryColumn({ type: 'numeric' })
+  @PrimaryColumn({ type: 'smallint' })
   role_id: number;
 
   @Column({ type: 'text', unique: true })
